+++ src/main/java/de/jeff_media/Drop2InventoryPlus/commands/CommandMain.java
@@ -2,6 +2,7 @@ package de.jeff_media.Drop2InventoryPlus.commands;
 
 import de.jeff_media.Drop2InventoryPlus.Config;
 import de.jeff_media.Drop2InventoryPlus.Main;
+import de.jeff_media.Drop2InventoryPlus.Messages;
 import de.jeff_media.Drop2InventoryPlus.Permissions;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
@@ -36,40 +37,40 @@ public class CommandMain implements CommandExecutor {
 		if(args.length>0 && sender.hasPermission(Permissions.ALLOW_TOGGLE_OTHERS)) {
 			Player player = Bukkit.getPlayer(args[0]);
 			if(player == null) {
-				sender.sendMessage(ChatColor.RED+"Player "+ChatColor.DARK_RED+args[0]+ChatColor.RED+" not found.");
+				Messages.sendMessage(sender,ChatColor.RED+"Player "+ChatColor.DARK_RED+args[0]+ChatColor.RED+" not found.");
 				return true;
 			}
 			main.getPlayerSetting(player).enabled=!main.getPlayerSetting(player).enabled;
 			if(main.getPlayerSetting(player).enabled) {
-				sender.sendMessage("§7Automatic drop collection has been §aenabled§7 for player "+player.getDisplayName());
+				Messages.sendMessage(sender,"§7Automatic drop collection has been §aenabled§7 for player "+player.getDisplayName());
 			} else {
-				sender.sendMessage("§7Automatic drop collection has been §cdisabled§7 for player "+player.getDisplayName());
+				Messages.sendMessage(sender,"§7Automatic drop collection has been §cdisabled§7 for player "+player.getDisplayName());
 			}
 			return true;
 		}
 
 		if(!(sender instanceof Player)) {
-			sender.sendMessage("You must be a player to run this command.");
+			Messages.sendMessage(sender,"You must be a player to run this command.");
 			return true;
 		}
 
 		Player p = (Player) sender;
 
 		if(!sender.hasPermission(Permissions.ALLOW_USE)) {
-			sender.sendMessage(main.getCommand("drop2inventory").getPermissionMessage());
+			Messages.sendMessage(sender,main.getCommand("drop2inventory").getPermissionMessage());
 			return true;
 		}
 
 		if(main.getConfig().getBoolean(Config.ALWAYS_ENABLED)) {
-			sender.sendMessage(ChatColor.RED+"Drop2Inventory cannot be disabled.");
+			Messages.sendMessage(sender,ChatColor.RED+"Drop2Inventory cannot be disabled.");
 			return true;
 		}
 		
 		main.togglePlayerSetting(p);
 		if(main.getPlayerSetting(p).enabled) {
-			sender.sendMessage(main.messages.MSG_ENABLED);
+			Messages.sendMessage(sender,main.messages.MSG_ENABLED);
 		} else {
-			sender.sendMessage(main.messages.MSG_DISABLED);
+			Messages.sendMessage(sender,main.messages.MSG_DISABLED);
 		}
 		return true;