+++ src/main/java/de/jeff_media/Drop2InventoryPlus/commands/CommandReload.java
@@ -1,6 +1,7 @@
 package de.jeff_media.Drop2InventoryPlus.commands;
 
 import de.jeff_media.Drop2InventoryPlus.Main;
+import de.jeff_media.Drop2InventoryPlus.Messages;
 import de.jeff_media.Drop2InventoryPlus.Permissions;
 import org.bukkit.ChatColor;
 import org.bukkit.command.Command;
@@ -11,12 +12,12 @@ public class CommandReload {
     static boolean run(Main main, CommandSender sender, Command command, String[] args)  {
 
         if(!sender.hasPermission(Permissions.ALLOW_RELOAD)) {
-            sender.sendMessage(command.getPermissionMessage());
+            Messages.sendMessage(sender,command.getPermissionMessage());
             return true;
         }
 
         main.reload();
-        sender.sendMessage(ChatColor.GREEN+"Drop2InventoryPlus has been reloaded.");
+        Messages.sendMessage(sender,ChatColor.GREEN+"Drop2InventoryPlus has been reloaded.");
         return true;
     }