+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.5.0
+- You can set messages to an empty String ("") to avoid them from being shown to the player
+
+## 1.4.1
+- Added config option "call-entity-pickup-item-event"
+
 ## 1.4.0
 - Drop2Inventory can now call fake EntityPickupItemEvents to check if other plugins want to cancel a drop picked up by the legacy drop detection
 - Fixed exception when other plugins somehow cause a EntityDamageEvent without giving a DamageCause